requirements.txt
lxml==4.7.1
requests==2.28.1
beautifulsoup4==4.11.1
selenium==4.7.2
django==4.0.3
djangorestframework==3.13.1
python-dotenv==0.14.0
psycopg2-binary==2.8.5
pandas==1.5.2